% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adf.r
\name{adf.apply}
\alias{adf.apply}
\title{Low level function for applying over an abstract data frame}
\usage{
adf.apply(x, FUN, type = c("data.frame", "model", "sparse.model"), formula,
  contrasts = NULL, subset = NULL, weights = NULL, na.action = NULL,
  offset = NULL, ..., chunk.max.line = 65536L, CH.MAX.SIZE = 33554432L,
  CH.MERGE = list, parallel = 1L)
}
\arguments{
\item{x}{an abstract data.frame}

\item{FUN}{function to apply over each chunk}

\item{type}{type of data to give as an input to FUN. If model
or sparse model, this is a list giving the response (y),
model matrix (x), weights (w), and offset (offset)
from the input forumal.}

\item{formula}{a formula to use with type equal to model or sparse.model}

\item{contrasts}{contrasts to use with type equal to model or sparse.model}

\item{subset}{a string to to use with type equal to model or sparse.model.
Will be evaluated in the environment of the data frame
(ex. subset="V2 + V3 > V4")}

\item{weights}{a string to to use with type equal to model or sparse.model.
Will be evaluated in the environment of the data frame.}

\item{na.action}{a function which indicates what should happen when the data
contain 'NA's. See lm.fit for more details.}

\item{offset}{a string to to use with type equal to model or sparse.model.
Will be evaluated in the environment of the data frame.}

\item{...}{Other arguments to pass.}

\item{chunk.max.line}{integer. Maximum number of lines to parse in a single chunk.}

\item{CH.MAX.SIZE}{integer. Maximum size in bytes of a chunk.}

\item{CH.MERGE}{method for combining the output of chunks. Typically list or rbind.}

\item{parallel}{integer. the number of parallel processes to use in the calculation
                        (*nix only).}
}
\value{
an object of class type
}
\description{
Low level function for applying over an abstract data frame
}

