% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adf.r
\name{adf}
\alias{adf}
\title{Create an abstract data frame}
\usage{
adf(description, conMethod = c("file", "gzfile", "bzfile", "xzfile"),
  sep = "|", nsep = NA, strict = TRUE, header = FALSE, colNames,
  colClasses, levels = list(), skip = 0L, nrowsClasses = 100L,
  chunkProcessor = identity, chunkFormatter, minSplits = 2L)
}
\arguments{
\item{description}{character string.  A description of the connection; the}

\item{conMethod}{string indicating the connection method.}

\item{sep}{character seperating the data columns. Ignored when
chunkFormatter is given.}

\item{nsep}{character seperating the column of rownames. Set to
NA to generate automatic rownames. Ignored when
chunkFormatter is given.}

\item{strict}{logical. Whether the parser should run in strict mode.
Ignored when chunkFormatter is given.}

\item{header}{logical indicating whether the first line of data, after
skip if provided, contains variable names. Ignored if
chunkFormatter is also provided, as adf would not know how
to parse this first row.}

\item{colNames}{an optional character vector of column names. If missing
and header=TRUE, these will be determined be the first
row of data. Otherwise, names will be constructed by
pasting the character 'V' with the column number.}

\item{colClasses}{an optional character vector of column classes. If
named and colNames is missing, the names will be
used for colNames. If missing, will be automatically
determined.}

\item{levels}{a named list, with names coorisponding to the colNames
of class character (or factor). Each element gives the
levels for the coorisponding variable. These will be
automatically determined when missing.}

\item{skip}{number of lines to strip off of the connection before
parsing}

\item{nrowsClasses}{number of rows to pull for determining colClasses and
factor levels when needed}

\item{chunkProcessor}{a function to apply post-processing to a formatted
dataframe. Usually only used without a chunkFormatter.}

\item{chunkFormatter}{an optional function turning the raw connection into
a dataframe. It must accept four parameters: data,
colNames, colClasses, levels. If missing this will
be constructed automatically.}

\item{minSplits}{an optional paramater for use with SparkR.}
}
\value{
An abstract data frame object, which can be used with iolm, ioglm, and
        other related functions.
}
\description{
Create an abstract data frame
}

