% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adf.r
\name{allFactorLevels}
\alias{allFactorLevels}
\title{Update abstract data frame with all factor levels}
\usage{
allFactorLevels(x)
}
\arguments{
\item{x}{an abstract dataframe}
}
\value{
a new abstract data frame, with (possibly) new
        factor levels derived from scanning the entire
        dataset.
}
\description{
Update abstract data frame with all factor levels
}

