% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adf.r
\name{as.adf}
\alias{as.adf}
\title{Convert a data.frame to an abstract data frame}
\usage{
as.adf(x, ...)
}
\arguments{
\item{x}{a data.frame}

\item{...}{other arguments. Currently unused}
}
\value{
an abstract data frame object
}
\description{
Convert a data.frame to an abstract data frame
}

