% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lars.r
\name{blars}
\alias{blars}
\title{Perform a big lasso or stepwise regression}
\usage{
blars(formula, data, data_frame_preprocessor = function(x) x,
  type = c("lasso", "lar", "forward.stagewise", "stepwise"),
  contrasts = NULL, sep = ",", eps = .Machine$double.eps,
  max.steps = NULL)
}
\arguments{
\item{formula}{the formual for the regression}

\item{data}{a connection to read data from}

\item{data_frame_preprocessor}{a function that turns the data read from
a connection into a properly formatted data frame}

\item{type}{One of "lasso", "lar", "forward.stagewise" or "stepwise". The
 names can be abbreviated to any unique substring. Default is
"lasso".}

\item{contrasts}{the contrasts for categorical regressors}

\item{sep}{if using a connection or file, which character is used as a separator between elements?}

\item{eps}{An effective zero}

\item{max.steps}{Limit the number of steps taken}
}
\description{
Perform a big lasso or stepwise regression
}

