% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ols.r
\name{blm}
\alias{blm}
\title{Perform a big linear regression}
\usage{
blm(formula, data, data_frame_preprocessor = function(x) x,
  contrasts = NULL, parallel = 1, tol = -1)
}
\arguments{
\item{formula}{the formual for the regression}

\item{data}{a connection to read data from}

\item{data_frame_preprocessor}{a function that turns the data read from
a connection into a properly formatted data frame}

\item{contrasts}{the contrasts for categorical regressors}

\item{parallel}{the number of cores to apply to the regression}
}
\description{
Perform a big linear regression
}

