% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lars.r
\name{coef.iolars}
\alias{coef.iolars}
\title{Get coefficients of an iolars object}
\usage{
\method{coef}{iolars}(object, ...)
}
\arguments{
\item{object}{an iolars object to get the coefficients matrix for}

\item{...}{other inputs; currently unused}
}
\description{
Get coefficients of an iolars object
}

