% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adf.r
\name{default.chunkFormatter}
\alias{default.chunkFormatter}
\title{Construct a default chunkFormatter}
\usage{
default.chunkFormatter(sep = sep, nsep = nsep, strict = strict)
}
\arguments{
\item{sep}{character seperating the data columns.}

\item{nsep}{character seperating the column of rownames. Set to
NA to generate automatic rownames.}

\item{strict}{logical. Whether the parser should run in strict mode.}
}
\description{
Construct a default chunkFormatter
}

