% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glm.r
\name{ioglm}
\alias{ioglm}
\title{Perform a generalized linear regression}
\usage{
ioglm(form, family = gaussian(), data, dfpp, beta_start = NULL,
  control = list(maxit = 25, epsilon = 1e-08, trace = FALSE),
  method = "irls", contrasts = NULL, sep = ",", parallel = 1, ...)
}
\arguments{
\item{form}{an object of class ‘"formula"’ (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{family}{a description of the error distribution and link function to be used in the model. This can be a character string naming a family function, a family function or the result of a call to a family function.}

\item{data}{a connection to read data from}

\item{dfpp}{the data frame preprocessor. a function that turns the
read data from a chunk into a properly formatted data frame}

\item{beta_start}{starting values for the linear predictor.}

\item{control}{a list of parameters for controlling the fitting process.}

\item{method}{the method to be used in fitting the model.  The default is
iteratively reweighted least squares (irls). Support for stochastic gradient
decent is coming.}

\item{contrasts}{an optional list. See the ‘contrasts.arg’ of ‘model.matrix.default’.}

\item{sep}{if using a connection or file, which character is used as a separator between elements?}

\item{parallel}{how many logical processor cores to use (default 1)}
}
\description{
Perform a generalized linear regression
}

