% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pcr.r
\name{iolm.pcr}
\alias{iolm.pcr}
\title{Recalculate iolm with Principal Component Regression}
\usage{
iolm.pcr(object, k = seq(1, ncol(object$xtx) - attr(object$terms,
  "intercept")), normalize = TRUE)
}
\arguments{
\item{object}{an iolm object with the desired formula}

\item{k}{A scalar or vector of degrees to fit.}

\item{normalize}{If TRUE, each variable is standardized to
have unit L2 norm, otherwise it is left alone.
Default is TRUE.}
}
\description{
This function takes a pre-computed iolm object,
and recalculates the regression vector with only
the top k-principal components.
}

