% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.r
\name{iolm}
\alias{iolm}
\title{Perform a linear regression}
\usage{
iolm(formula, data, subset = NULL, weights = NULL, na.action = NULL,
  offset = NULL, contrasts = NULL, tol = -1, parallel = 1L)
}
\arguments{
\item{formula}{the formula for the regression}

\item{data}{an abstract data frame, or something which can be
coerced to one.}

\item{subset}{an options character string, which will be evaluated in the
frame of the data, to indicate which rows to include
in the analysis}

\item{weights}{a optional character string, which will be evaluated in the
frame of the data, giving the sample weights for the regression}

\item{na.action}{a function which indicates what should happen when the data
contain 'NA's. See lm.fit for more details.}

\item{offset}{a optional character string, which will be evaluated in the
frame of the data, giving the offsets for the regression}

\item{contrasts}{contrasts to use with the regression. See the \code{contrasts.arg}
of \code{model.matrix.default}}

\item{tol}{numeric tolerance. Set to -1 to ignore.}

\item{parallel}{integer. the number of parallel processes to use in the
calculation (*nix only).}
}
\description{
Perform a linear regression
}

