% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.r
\name{iolm.ridge}
\alias{iolm.ridge}
\title{Add a ridge penalty to an iolm object}
\usage{
iolm.ridge(object, lambda = seq(0, 1, by = 0.1), normalize = TRUE)
}
\arguments{
\item{object}{an iolm object with the desired formula}

\item{lambda}{A scalar or vector of ridge constants.}

\item{normalize}{If TRUE, each variable is standardized to have unit L2 norm,
otherwise it is left alone. Default is TRUE.}
}
\description{
This function takes a pre-computed iolm object,
and applies a (sequence) of ridge penalties. An
intercept, if present in the original model, will
not be penalized.
}

