% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lmnet.r
\name{iolmnet}
\alias{iolmnet}
\title{Fit a linear model with lasso or elasticnet regularization}
\usage{
iolmnet(formula, data, subset = NULL, weights = NULL, na.action = NULL,
  offset = NULL, alpha = 1, lambda = NULL, contrasts = NULL,
  standardize = FALSE, tol = 1e-07, max_it = 1e+05, filter = c("strong",
  "safe", "none"), lambda_epsilon = 1e-04, nlambdas = 100, parallel = 1L)
}
\arguments{
\item{formula}{the formulat for the regression}

\item{data}{an abstract data frame or something that can be coerced into one.}

\item{subset}{an options character string, which will be evaluated in the
frame of the data, to indicate which rows to include in the analysis}

\item{weights}{a optional character string, which will be evaluated in the
frame of the data, giving the sample weights for the
regression.}

\item{na.action}{a function which indicates what should happen when the data
contain 'NA's. See lm.fit for more details.}

\item{offset}{a optional character string, which will be evaluated in the
frame of the data, giving the offsets for the regression}

\item{alpha}{the elasticnet mixing parameter 0 <= alpha <= 1.}

\item{lambda}{a user supplied value (or sequence of values) for the penalty
parameter. If not specified then a regularization path is created based
on the data.}

\item{contrasts}{contrasts to use with the regression. See the
\code{contrasts.arg} of \code{model.matrix.default}}

\item{standardize}{should the regression variables be normalized to have
mean zero and standard deviation one?}

\item{tol}{numeric tolerance.}

\item{max_it}{the maximum number of iterations per regression.}

\item{filter}{should filtering rules be used to remove variables that will
not be needed in the regression? Default is strong corresponding to
http://www-stat.stanford.edu/~tibs/ftp/strong.pdf.}

\item{lambda_epsilon}{this value is multiplied by the max lambda in the
data to determine the minimum lambda when the regularization path is
determined from the data. This is ignored when lambda is specified.}

\item{nlambdas}{the number of lambdas to be generated in the regularization
path. This is ignored when lambda is specified.}

\item{parallel}{integer. the number of parallel processes to use in the
calculation (*nix only).}
}
\description{
Fit a linear model via penalized maxiumum likelihood.
The regularization path is computed for the lasso or elasticnet
penalty at a grid of values for the regularization parameter
lambda. Can deal data frames or abstract data frames.
}

