% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lp.r
\name{iolp}
\alias{iolp}
\title{Fit a robust linear regression}
\usage{
iolp(formula, data, p = 1.5, weights = NULL, subset = NULL,
  na.action = NULL, offset = NULL, contrasts = NULL, beta_init = NULL,
  delta = 1e-04, maxit = 20, acc = 1e-05, tol = -1)
}
\arguments{
\item{formula}{the formula for the regression}

\item{data}{an abstract data frame, or something which can be
coerced to one.}

\item{weights}{a optional character string, which will be evaluated in the
frame of the data, giving the sample weights for the regression}

\item{subset}{an options character string, which will be evaluated in the
frame of the data, to indicate which rows to include
in the analysis}

\item{na.action}{a function which indicates what should happen when the data
contain 'NA's. See lm.fit for more details.}

\item{offset}{a optional character string, which will be evaluated in the
frame of the data, giving the offsets for the regression}

\item{contrasts}{contrasts to use with the regression. See the \code{contrasts.arg}
of \code{model.matrix.default}}

\item{maxit}{the limit on the number of IWLS iterations.}

\item{acc}{accuracy for the IWLS stopping criterion.}

\item{tol}{numeric tolerance. Set to -1 to ignore.}

\item{a}{tuning parameter for Tukey bisquare. See Details for more information.}
}
\description{
Fits an M-estimator using Tukey's bisquare function for
estimating slope coefficients in a linear model.
}
\details{
The parameter \code{a} controls the tradeoff between efficency
and robustness. The default value of 4.685 yields an efficency of 95%
but breakdown point of about 10%; conversely 1.547 has only a 28% efficency
but a breakdown point of 50%. For large datasets, it may be preferable to set
the values of \code{a} lower than the default.
}

