\name{ioregression-package}
\alias{ioregression-package}
\alias{ioregression}
\docType{package}
\title{
Out-of-core regressions with connection inputs
}
\description{
This package implements out-of-core regression routines (OLS, GLM,
LARS) and can be used with any input implemented as an R
connection.
}
\details{
\tabular{ll}{
Package: \tab ioregression\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-01-22\cr
License: \tab Apache License 2.0\cr
}

Index of functions:
\preformatted{
iolm
summary.lm
iolars
}
}
\author{
Taylor Arnold, Michael J. Kane, Simon Urbanek

Maintainer: Michael J. Kane <kaneplusplus@gmail.com>
}
\examples{
runif(10)
}
