% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glm.r
\name{print.ioglm}
\alias{print.ioglm}
\title{Print ioglm object}
\usage{
\method{print}{ioglm}(x, ...)
}
\arguments{
\item{x}{output of iolm}

\item{...}{optional arguments passed to print.glm}
}
\description{
Print ioglm object
}

