% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lars.r
\name{print.iolars}
\alias{print.iolars}
\title{Print an iolars object}
\usage{
\method{print}{iolars}(x, ...)
}
\arguments{
\item{x}{an iolars object to print the summary of}

\item{...}{other inputs; currently unused}
}
\description{
Print an iolars object
}

