% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pcr.r
\name{print.iolm.pcr}
\alias{print.iolm.pcr}
\title{Print an iolm.pcr object}
\usage{
\method{print}{iolm.pcr}(x, ...)
}
\arguments{
\item{x}{an iolars object to print the summary of}

\item{...}{other inputs; currently unused}
}
\description{
Print an iolm.pcr object
}

