% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ols.r
\name{print.iolm}
\alias{print.iolm}
\title{Print iolm object}
\usage{
\method{print}{iolm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{output of iolm}

\item{digits}{significant digits to print}

\item{...}{optional, currently unused, arguments}
}
\description{
Print iolm object
}

