% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/io.pcr.r, R/ridge.r
\name{print.iolm.ridge}
\alias{print.iolm.ridge}
\title{Print an ioridge object}
\usage{
\method{print}{iolm.ridge}(x, ...)

\method{print}{iolm.ridge}(x, ...)
}
\arguments{
\item{x}{an iolars object to print the summary of}

\item{...}{other inputs; currently unused}

\item{x}{an iolars object to print the summary of}

\item{...}{other inputs; currently unused}
}
\description{
Print an ioridge object

Print an ioridge object
}

