% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lp.r
\name{print.iolp}
\alias{print.iolp}
\title{Print iolp object}
\usage{
\method{print}{iolp}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{output of iolp}

\item{digits}{significant digits to print}

\item{...}{optional, currently unused, arguments}
}
\description{
Print iolp object
}

