% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlm.r
\name{print.iorlm}
\alias{print.iorlm}
\title{Print iorlm object}
\usage{
\method{print}{iorlm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{output of iorlm}

\item{digits}{significant digits to print}

\item{...}{optional, currently unused, arguments}
}
\description{
Print iorlm object
}

