% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ols.r
\name{print.summary.iolm}
\alias{print.summary.iolm}
\title{Print iolm summary}
\usage{
\method{print}{summary.iolm}(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{output of summary.iolm}

\item{digits}{significant digits to print}

\item{symbolic.cor}{logical. Should symbolic correlation be printed.}

\item{signif.stars}{logical. Should significant starts be printed.}

\item{...}{optional, currently unused, arguments}
}
\description{
Print iolm summary
}

