% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ols.r
\name{summary.iolm}
\alias{summary.iolm}
\title{Get the regression diagnostics for a linear regression}
\usage{
\method{summary}{iolm}(object, ...)
}
\arguments{
\item{object}{an object return from iolm}

\item{...}{optional, currently unused, arguments}
}
\description{
Get the regression diagnostics for a linear regression
}

