% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glm.r
\name{dfpp_gen}
\alias{dfpp_gen}
\title{Data frame preprocessor}
\usage{
dfpp_gen(col_names, col_types, sep, dfpp_fun = function(x) x)
}
\arguments{
\item{col_types}{the column types of the data}

\item{sep}{the column separator}

\item{dfpp_fun}{the data.frame preprocessing function}
}
\description{
Data frame preprocessor
}

