% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm.r
\name{ioglm}
\alias{ioglm}
\title{Perform a generalized linear regression}
\usage{
ioglm(formula, family = gaussian, data, weights = NULL, subset = NULL,
  na.action = NULL, start = NULL, etastart, mustart, offset = NULL,
  control = list(), contrasts = NULL, tol = -1)
}
\arguments{
\item{formula}{the formula for the regression}

\item{family}{a description of the error distribution and link function to
be used in the model. This can be a character string naming a
family function, a family function or the result of a call to
a family function.}

\item{data}{an abstract data frame, or something which can be
coerced to one.}

\item{weights}{a optional character string, which will be evaluated in the
frame of the data, giving the sample weights for the regression}

\item{subset}{an options character string, which will be evaluated in the
frame of the data, to indicate which rows to include
in the analysis}

\item{na.action}{a function which indicates what should happen when the data
contain 'NA's. See lm.fit for more details.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{etastart}{starting values for the linear predictor.}

\item{mustart}{starting values for the vector of means.}

\item{offset}{a optional character string, which will be evaluated in the
frame of the data, giving the offsets for the regression}

\item{control}{a list of parameters for controlling the fitting process.}

\item{contrasts}{contrasts to use with the regression. See the \code{contrasts.arg}
of \code{model.matrix.default}}

\item{tol}{numeric tolerance. Set to -1 to ignore.}
}
\description{
Perform a generalized linear regression
}

