% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ols.r
\name{iolm}
\alias{iolm}
\title{Perform a linear regression}
\usage{
iolm(formula, data, data_frame_preprocessor = function(x) x,
  contrasts = NULL, tol = -1, sep = ",")
}
\arguments{
\item{formula}{the formual for the regression}

\item{data}{a connection to read data from}

\item{data_frame_preprocessor}{a function that turns the data read from
a connection into a properly formatted data frame}

\item{contrasts}{the contrasts for categorical regressors}

\item{sep}{if using a connection or file, which character is used as a separator between elements?}
}
\description{
Perform a linear regression
}

