% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ols.r
\name{iolm}
\alias{iolm}
\title{Perform a linear regression}
\usage{
iolm(form, data, dfpp, contrasts = NULL, tol = -1, sep = ",",
  parallel = 1)
}
\arguments{
\item{form}{the formula for the regression}

\item{data}{a connection to read data from}

\item{dfpp}{a function that turns the data read from
a connection into a properly formatted data frame}

\item{contrasts}{the contrasts for categorical regressors}

\item{sep}{if using a connection or file, which character is used as a separator between elements?}

\item{parallel}{how many logical processor cores to use (default 1)}
}
\description{
Perform a linear regression
}

