% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.r
\name{iolp}
\alias{iolp}
\title{Fit an Lp Regression}
\usage{
iolp(formula, data, p = 1.5, weights = NULL, subset = NULL,
  na.action = NULL, offset = NULL, contrasts = NULL, beta_init = NULL,
  delta = 1e-04, maxit = 20, acc = 1e-05, trace = FALSE, tol = -1)
}
\arguments{
\item{formula}{the formula for the regression}

\item{data}{an abstract data frame, or something which can be
coerced to one.}

\item{p}{value of p for the regression, such that 1 <= p < 2.
Using p=2 is also allowed for testing purposes, but
sub-optimal as iolm should instead be called directly.}

\item{weights}{a optional character string, which will be evaluated in the
frame of the data, giving the sample weights for the regression}

\item{subset}{an options character string, which will be evaluated in the
frame of the data, to indicate which rows to include
in the analysis}

\item{na.action}{a function which indicates what should happen when the data
contain 'NA's. See lm.fit for more details.}

\item{offset}{a optional character string, which will be evaluated in the
frame of the data, giving the offsets for the regression}

\item{contrasts}{contrasts to use with the regression. See the \code{contrasts.arg}
of \code{model.matrix.default}}

\item{beta_init}{initial beta vector to start at; when missing an first pass using
iolm is run to determine the starting point.}

\item{delta}{tuning parameter to provide weights for very small residuals.
most users should not need to change this parameter.}

\item{maxit}{the limit on the number of IWLS iterations.}

\item{acc}{accuracy for the IWLS stopping criterion.}

\item{trace}{logical indicating if output should be produced for each
iteration.}

\item{tol}{numeric tolerance. Set to -1 to ignore.}
}
\description{
Minimizes the loss ||Y - Xb||_p, for a suitable choice
of p.
}

