% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lars.r
\name{lars_par}
\alias{lars_par}
\title{This is a modified version of the lars function from the lars package,
which allows for pre-computing XtY.}
\usage{
lars_par(XtX, XtY, type = c("lasso", "lar", "forward.stagewise", "stepwise"),
  eps = .Machine$double.eps, max.steps = NULL)
}
\arguments{
\item{XtX}{precomputed XtX matrix}

\item{XtY}{precomputed XtY vector}

\item{type}{One of "lasso", "lar", "forward.stagewise" or "stepwise". The
names can be abbreviated to any unique substring. Default is
"lasso".}

\item{eps}{An effective zero}

\item{max.steps}{Limit the number of steps taken;}
}
\description{
This is a modified version of the lars function from the lars package,
which allows for pre-computing XtY.
}

