% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adf.r
\name{lm.model.frame}
\alias{lm.model.frame}
\title{A custom version of lm to construct a model frame that
does not drop unused levels (because these should be)
in other chunks.}
\usage{
lm.model.frame(formula, data, subset, weights, na.action, method = "qr",
  model = TRUE, x = FALSE, y = FALSE, qr = TRUE, singular.ok = TRUE,
  contrasts = NULL, offset, ...)
}
\description{
A custom version of lm to construct a model frame that
does not drop unused levels (because these should be)
in other chunks.
}

