% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adf.r
\name{print.adf}
\alias{print.adf}
\title{Print the details of an abstract data frame}
\usage{
\method{print}{adf}(x, ...)
}
\arguments{
\item{x}{a data.frame}

\item{...}{other arguments. Currently unused}
}
\description{
Print the details of an abstract data frame
}

