% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm.r
\name{summary.ioglm}
\alias{summary.ioglm}
\title{Get the regression diagnostics for a linear regression}
\usage{
\method{summary}{ioglm}(object, ...)
}
\arguments{
\item{object}{an object return from ioglm}

\item{...}{optional, currently unused, arguments}
}
\description{
Get the regression diagnostics for a linear regression
}

