% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/glm.r
\name{summary.ioglm}
\alias{summary.ioglm}
\title{Get the regression diagnostics for a linear regression}
\usage{
\method{summary}{ioglm}(object, data, parallel = 1, ...)
}
\arguments{
\item{object}{an object return from ioglm}

\item{data}{a data.frame or connection to the data set where training was performed.}

\item{parallel}{how many logical processor cores to use (default 1)}

\item{...}{optional, currently unused, arguments}
}
\description{
Get the regression diagnostics for a linear regression
}

