% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ols.r
\name{summary.iolm}
\alias{summary.iolm}
\title{Get the regression diagnostics for a linear regression}
\usage{
\method{summary}{iolm}(object, parallel, data, ...)
}
\arguments{
\item{object}{an object return from iolm}

\item{parallel}{how many logical processor cores to use (default 1)}
}
\description{
Get the regression diagnostics for a linear regression
}

