% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ridge.r
\name{summary.iolm.ridge}
\alias{summary.iolm.ridge}
\title{Summarize a ridge regression for a particular value of lambda}
\usage{
\method{summary}{iolm.ridge}(object,
  lambda = object$lambda[which.min(object$AIC)], ...)
}
\arguments{
\item{object}{an iolars object to print the summary of}

\item{lambda}{a single lambda value}

\item{...}{other inputs; currently unused}
}
\description{
Summarize a ridge regression for a particular value of lambda
}

